+++ src/Extension/Mention/MentionExtension.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace League\CommonMark\Extension\Mention;
 
 use League\CommonMark\Environment\ConfigurableEnvironmentInterface;
+use League\CommonMark\Exception\InvalidOptionException;
 use League\CommonMark\Extension\ExtensionInterface;
 use League\CommonMark\Extension\Mention\Generator\MentionGeneratorInterface;
 
@@ -25,8 +26,12 @@ final class MentionExtension implements ExtensionInterface
         foreach ($mentions as $name => $mention) {
             foreach (['prefix', 'regex', 'generator'] as $key) {
                 if (! \array_key_exists($key, $mention)) {
-                    throw new \RuntimeException(\sprintf('Missing "%s" from MentionParser configuration', $key));
+                    throw new InvalidOptionException(\sprintf('Required option "mentions/%s/%s" for the Mention extension is missing', $name, $key));
                 }
+            }
+
+            if (! self::isAValidPartialRegex($mention['regex'])) {
+                throw InvalidOptionException::forConfigOption(\sprintf('mentions/%s/regex', $name), $mention['regex'], 'Invalid partial regex. Make sure to exclude starting/ending delimiters and flags.');
             }
 
             if ($mention['generator'] instanceof MentionGeneratorInterface) {
@@ -36,8 +41,15 @@ final class MentionExtension implements ExtensionInterface
             } elseif (\is_callable($mention['generator'])) {
                 $environment->addInlineParser(MentionParser::createWithCallback($mention['prefix'], $mention['regex'], $mention['generator']));
             } else {
-                throw new \RuntimeException(\sprintf('The "generator" provided for the MentionParser configuration must be a string template, callable, or an object that implements %s.', MentionGeneratorInterface::class));
+                throw new InvalidOptionException(\sprintf('The "generator" provided for the "%s" MentionParser configuration must be a string template, callable, or an object that implements %s.', $name, MentionGeneratorInterface::class));
             }
         }
+    }
+
+    private static function isAValidPartialRegex(string $regex): bool
+    {
+        $regex = '/' . $regex . '/i';
+
+        return @\preg_match($regex, '') !== false;
     }
 }