+++ tests/functional/Extension/Mention/MentionExtensionTest.php
@@ -15,6 +15,7 @@ namespace League\CommonMark\Tests\Functional\Extension\Mention;
 
 use League\CommonMark\CommonMarkConverter;
 use League\CommonMark\Environment\Environment;
+use League\CommonMark\Exception\InvalidOptionException;
 use League\CommonMark\Extension\Mention\Generator\MentionGeneratorInterface;
 use League\CommonMark\Extension\Mention\Mention;
 use League\CommonMark\Extension\Mention\MentionExtension;
@@ -139,7 +140,7 @@ EOT;
 
     public function testConfigUnknownGenerator(): void
     {
-        $this->expectException(\RuntimeException::class);
+        $this->expectException(InvalidOptionException::class);
 
         $environment = Environment::createCommonMarkEnvironment();
         $environment->addExtension(new MentionExtension());
@@ -160,7 +161,7 @@ EOT;
 
     public function testLegacySymbolOption(): void
     {
-        $this->expectException(\RuntimeException::class);
+        $this->expectException(InvalidOptionException::class);
 
         $environment = Environment::createCommonMarkEnvironment();
         $environment->addExtension(new MentionExtension());
@@ -169,6 +170,27 @@ EOT;
                 'github_handle' => [
                     'symbol'    => '@',
                     'regex'     => '[a-z\d](?:[a-z\d]|-(?=[a-z\d])){0,38}(?!\w)',
+                    'generator' => 'https://github.com/%s',
+                ],
+            ],
+        ]);
+
+        $converter = new CommonMarkConverter([], $environment);
+
+        $converter->convertToHtml('foo');
+    }
+
+    public function testWithFullRegexOption(): void
+    {
+        $this->expectException(InvalidOptionException::class);
+
+        $environment = Environment::createCommonMarkEnvironment();
+        $environment->addExtension(new MentionExtension());
+        $environment->setConfig([
+            'mentions' => [
+                'github_handle' => [
+                    'prefix'    => '@',
+                    'regex'     => '/[a-z\d](?:[a-z\d]|-(?=[a-z\d])){0,38}(?!\w)/i',
                     'generator' => 'https://github.com/%s',
                 ],
             ],